Show a banner in QS for EOL devices
The banner is shown as part of system info (header with QsRevamp
enabled) in quick settings, the same way as channel indicator.
It gets shown when the device reaches its end of life (as reported
by the update engine). This leaves a potential conflict in case both
EOL notice and channel indicator have to be shown - to handle this case
channel indicator gets hidden when EOL notice is visible.

BUG=b:272319343

Change-Id: I6b8ef350026f295ec483545382a5e9dab2c3d01e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4322087
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119615}