media/gpu/v4l2: restart output queue conditionally
When CAPTURE was not streaming before it might be
the case that it was not setup yet. That means
buffers were not allocated and STREAMON will fail.
Adding condition checking if input and output was
streaming before safes this situation.
This change removes also code duplication that it
introduced.

Bug: b:232394619
Bug: b:271159608
Test: run android.mediav2.cts.CodecDecoderSurfaceTest
Test: tast run 127.0.0.1:2255 video.ChromeStackDecoder.h264
Change-Id: I64961c48b4015075752b1a8d4bcd9b4e903f8ed4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4348454
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119646}