[iOS] Fix compile errors when bumping minimum iOS deployment target
Code in the else block for these methods are deprecated. Even though
we have an if(@available) statement, it doesn't stop the compile
warning when we bump the minimum iOS deployment target from 14 to 15
Therefore, we have added compile guards to avoid compiling code that
isn't needed in iOS builds. For mac builds, code should compile as-is.

After placing the compile guards, we get a non-void method error since the remaining return statements are in an if block, so we add a return
statement that should effectively do nothing as we should fall into
the other logic paths and "early" return.

Bug: 1406561
Change-Id: I6f3ab1c58f3ef1ebb35677370075ce9a53e46b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4347606
Commit-Queue: Matt Mueller <mattm@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Auto-Submit: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119647}