[Preloads] Check UrlLoaderFactory is not null before using.
In this CL we add a CHECK_IS_TEST call when the url_loader_factory is
null so that when it is used during a test that doesn't have
url_loader_factory set up, the test doesn't fail. This covers us for
tests not related to preloads that are using profile builder and don't
set up url_loader_factory.

Bug: b/273368914
Change-Id: I4be30cb70a88603b35f25baf52d47d7d175416f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4352368
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Auto-Submit: Melissa Zhang <melzhang@chromium.org>
Commit-Queue: Melissa Zhang <melzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119692}