Fix offset adjustment when computing MagnifierPaintFilter src_rect.
The current computation of the MagnifierPaintFilter src_rect doesn't
quite handle clipping correctly, which causes the ChromeOS magnifying
glass to sometimes not show the screen edges (see screen recordings in
crbug/1380209). Fix it based on this computation and diagram:
https://docs.google.com/document/d/1pwARyNTMWYf0N2FW3aoX0E3C_8cha56M6iJ3_OmQbGQ/edit?usp=sharing&resourcekey=0-c51EedcRMdxuw03qDZtw4A.

Note that the magnifier center should usually be on the screen, so the
center of the unclipped rectangle should be in the clipped rectangle.
However, when the magnifier center is very close to the edge of the
screen, it is sometimes computed as a few pixels outside the clipped
rectangle. In these cases, we take the zoom center to be the closest
point inside the clipped rectangle.

Also fix the clipping offset computed in SoftwareRenderer, which had
the wrong sign on the y value based on the diagrams and from the
corresponding computation in SkiaRenderer:
https://source.chromium.org/chromium/chromium/src/+/main:components/viz/service/display/skia_renderer.cc;l=2901;drc=5717f6af2d29e4b80088db0ca801552ba8bf60ed

Bug: 1380209
Change-Id: I1985e602fa75d196bdb9b7c48d7a8de788f08ea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4349714
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Cr-Commit-Position: refs/heads/main@{#1119699}