[Extensions] Handle window.close() for side panel
Calling window.close() from an extension side panel should not close the
browser window. Instead, it should close the side panel if the
extension's panel is shown (this invalidates the extension's side panel
view, and deletes the cached view if an entry is registered but not
currently shown.

Bug: 1378048
Change-Id: Ia4b47719fe3ae8bbbc1d223352a2493a614b4e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4307862
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119706}