[omnibox][ml] Ensures scoring model callback is always run when enabled
This change makes a few modifications to the current flow to ensure the
model and its callback are invoked when expected:

The scoring model is expected to run only once all the providers
are done. However currently it may also run when
`omnibox::kSingleSortAndCullPass` is disabled. This CL addresses
that.

The scording model is expected to run for URL matches only. However,
currently it is invoked for non-URL matches too. This CL ensures that
the model is run for URL matches only.

This CL also ensures the model callback is always invoked when expected
and updates some of the comments in the ACCotroller for more clarity.

Bug: 1405555
Change-Id: Idb1fb211018c5fc91aea17073d2a889968a381da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4354251
Reviewed-by: Angela Yoeurng <yoangela@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119724}