Add "fil" and "zh-HK" handwriting packs to LP manager
Two IMEs - Filipino and Cantonese - rely on these language packs to
exist. These two DLCs will be added in the CrOS CLs
https://crrev.com/c/4357696/ and https://crrev.com/c/4357697/.

Note that the contents of handwriting-fil and handwriting-tl are
identical, and the latter will be removed in a follow-up CL as it is
unused.

Bug: b:270264583
Change-Id: I9d7aa543ade3870f400ee73c16e79978eacd89fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4357417
Auto-Submit: Michael Cui <mlcui@google.com>
Commit-Queue: Michael Cui <mlcui@google.com>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119774}