[ios] fix FrameRateDeciderTest.OptimalFrameSinkIntervalIsPicked test
That test verifies that the FrameRateDecider chooses min refresh
interval using supported intervals the test sets. However,
those intervals are not used and ignored in case of iOS as the
system allows to set any preferred refresh rate and will adapt
accordingly. For example, if a display supports 60Hz, the supported
refresh rates (aka the rates that the system will choose) are 15, 20,
30, and 60 FPS respectively. And if a client sets a refresh rate
to be 28 FPS, it will be rounded to 30 by the system.

Cq-Include-Trybots: luci.chromium.try:ios-blink-dbg-fyi
Bug: 1413559,1421171
Change-Id: Icb775ba52b9a4388522297b0db6f16218e436c06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4352947
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1119802}