[signin] Add metrics to count users with previous sync accounts
PrefService holds the information about whether sync was turned on
previously on the same profile and later turned off (at least once).
This patch essentially logs the value of this preference (turned into
a boolean) upon profile startup, to get a sense of what the fraction
of users could be in a state where they carry over data from a
previous sync-enabled session.

The two proposed metrics are very similar but differ on the exact preconditions required to log the metric: one of them gets logged for
fully-signed-out users, whereas the other one is more relaxed and
logged also if the current user is signed-in-with-sync-off.

Change-Id: I53816daf6d5430ed70dfdce11ba52b0f61e0a639
Fixed: 1425443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4348462
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119899}