Separate Remote from Local Web approval
This refactoring is a pre-requisite for moving the Web
Approvals Manager source to components and removing
chrome/browser and content dependencies from it.
This change splits the local web approval functionality
that requires access to content into a separate class,
with platform-specific implementations (Web Content Handler).

The new Handler belongs to the interstitial, and more
content-specific functionality will be move there in
subsequent changes.

Bug: b/265758915
Change-Id: If97381da9856ba4f212ca389d98877d8c794896a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4334759
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Anthi Orfanou <anthie@google.com>
Cr-Commit-Position: refs/heads/main@{#1119915}