Replace calls to legacyCreateTypeface() in FontCacheSkia
... by matchFamilyStyle().

SkFontMgr::legacyCreateTypeface() isn't expected to be used in
Chrome/Blink code anymore, and now that matchFamilyStyle() is
available, use it instead.

Note that particularly on iOS, the proper behavior of the callee
method (FontCacheSkia::CreateTypeface()) relies on [1], which is
being worked out in Skia by bungeman@chromium.org in parallel.

[1] https://skia-review.googlesource.com/c/skia/+/657658

TEST=FontUpdateInvalidationTest.FallbackBetweenPendingAndLoadedCustomFonts
BUG=1424383
R=bungeman@chromium.org, vollick@chromium.org

Change-Id: Ic7f624f134383459f6d28050da83dbe3b54aab52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4346973
Reviewed-by: Ben Wagner <bungeman@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1119921}