Have android webview's goBack/goForward check canGoBack/canGoForward
Within the chromium implementation, it is expected that goBack/goForward
are only called when it is possible to go back/forward. The public
webview API does not appear to have this as a precondition. So we have
the webview goBack/goForward implementations check this.

Bug: 1425883
Change-Id: I7973594593e3e2ac6bc32c52145311096fa0899a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4353976
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119936}