Remove unused ChromeOS specific code out of Supervised User Service
This is part of removing the references of the profile_
object within Supervised User Service.
The theme service trigger (ChromeOS specific code) is no longer
required to be called: the code removed was used on switching
between supervised and unsupervised profiles, a use case that
does not exist anymore on ChromeOS.

This refactoring is a pre-requisite for moving the source
to components.

Bug: b/272008432
Change-Id: Iea91cbe2b914ff974aea636b8206178de52918c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4329789
Commit-Queue: Anthi Orfanou <anthie@google.com>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119968}