Allow limiting precision of bid and score for FLEDGE reporting
This CL modifies FLEDGE reporting so that the floating-point precision
of the bid and score passed to FLEDGE event-level reporting is limited.
By default the floating point exponent is limited to 8 bits, but the
mantissa is still allowed the full 53 bits. Eventually we will decrease
the precision further, to further limit the ability exfiltrate
identifying information from FLEDGE auctions.

Change-Id: Ic477ad41cc9ce5e66a7ce6a946bc30f0b0cfab2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4329459
Commit-Queue: Russ Hamilton <behamilton@google.com>
Reviewed-by: Qingxin Wu <qingxinwu@google.com>
Cr-Commit-Position: refs/heads/main@{#1120021}