[Adaptive Toolbar] Added controller for bookmark toolbar button
This CL adds a controller for a button to bookmark the current page.
This button will be shown on the Android toolbar next to the tab
switcher button.

This CL only adds the UI controller, a follow-up CL
(https://crrev.com/c/4348520) will hook it up
to the rest of the UI and add it to settings for users to enable.

Bug: 1425502
Change-Id: Ic06958b46e776d75f200c19e4be6681807f77667
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4354835
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Salvador Guerrero Ramos <salg@google.com>
Cr-Commit-Position: refs/heads/main@{#1120077}