Reset LocationBar translation X.
Reset LocationBar translation X as well. This was leaking translation
between pages and layouts, and when the NTP was loaded but not scrolled,
this could cause the UrlBar to draw overtop of the home button on the
next transition.

Typically the toolbar bitmap capture draws over top of this, so it's
difficult to see this glitch. But further changes will exacerbate this
if it is not fixed.

Bug: 1425438
Change-Id: I72082d0ec05a00df8ef451b6dd063753198affa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4357757
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120098}