Add Trust Token access observer.
This adds support for an observer whenever a Trust Token operation
occurs.

This hooks up the models in the network layer to the content layer, calling OnTrustTokensAccess callbacks when the API is triggered (primarily through network requests/navigations). Prior to being
used in these callbacks, the network mojo type is converted into a
content level TrustTokenAccessDetails type.

Users of this observer can extend WebContentsObserver to receive the
events and react accordingly. The primary use of this will be to hook
into the content/data model in the UX.


Bug: 1378264
Change-Id: I6e24415c6d1ba81029d24141277f276fd742917e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3969177
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Steven Valdez <svaldez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120101}