[rebaseline-cl] Add deprecation message for `--no-resultDB`
The ResultDB-enabled path has been stable for 2+ months now, so
`--no-resultDB` is no longer needed. Removing this flag will allow
for code cleanups.

No plans to remove this flag soon, and this flag is probably not used
much (it was not well-publicized). Still, we want to warn any users with
as much lead time as possible.

Bug: 1406660
Change-Id: I569e8dab4ed4b4cd62d6d7a9bb2f6a1c5f898889
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4358192
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1120105}