webauthn: replace the mechanism priority system.
The code for figuring out which mechanism to trigger immediately has
aged poorly. This change replaces it with a functionally equivalent,
but hopefully easier to read, version. Flag guarded: the old code still
exists and is A/B tested against the new.

This change is not intended to change any behaviour, only to set the
code on a cleaner path.

Change-Id: I003375515464b4cb1a7922855f825a95caf6e2cb
Fixed: 1426245
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4353713
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120106}