[User Education] Update anchor positions on body resize.
This fixes a bug where resizing the side panel in RTL would not update
the position of a floating help bubble. This adds the body of the
document to the list of things the help bubble mixin watches; if the
body resizes, element bounds are updated and (if they change) are sent
to the help bubble system to update floating help bubbles.

See attached bug for issue.

Bug: 1425487
Change-Id: I734c66ceec535685d1f92538751681c5b0e18fc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4355094
Reviewed-by: Mickey Burks <mickeyburks@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120122}