Centralize Session Logic in XrSessionCoordinator
Removes some logic and indirection that was happening in XrDelegateImpl
via ArDelegate and has XrDelegateImpl directly listen to
XrSessionCoordinator. Updates XrSessionCoordinator to better manage
and supply the state of the active session as well as its type. This
has the side effect of rendering the ArDelegate unnecessary, now that
the XrSessionCoordinator is always available to the XrDelegateImpl.
This also moves the SessionType enum to the XrSessionCoordinator which
is a more logical "source of truth". Note that much of this logic is
handled via statics that are private to the XrSessionCoordinator class
as each of the native runtimes will create their own
XrSessionCoordinator.

In preparation for some of the pending Vr logic being added to
XrSessionCoordinator, abstracts some useful shared logic to
startSession and renames the existing startSession method to
startArSession.

Bug: 989117
Change-Id: I744f2a0da3aa36bb2e1be622c641cf087eec5132
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4343101
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120125}