[NTP][Quests] Cleanup of quests flag and image check usage
* Updated quests module to not expect journeys or journeys images to be
  enabled. Instead it just checks that ImageService is enabled, and it
  checks for the entire module, instead of per image.
* Check isKnownToSync at the layout selection logic level, since we can
  not show those images anyways if that value is false.
* Change suggestImages to false, as requested.

For the binding changes:
Since we didn't want to depend on Journeys being enabled, I wanted to
move the binding out of the Journeys check. However, then Chrome would
crash if both Journeys and the Quests module were enabled at the same
time. There's a crash report
(https://bugs.chromium.org/p/chromium/issues/detail?id=1424935) that
suggests that the same thing was happening for bookmarks. To deal with
this, I just checked if any of the UIs that use ImageService are
enabled and bind it to all of them.

Bug: 1425245, 1426169, 1424935
Change-Id: I717f040085105d23c6fe9fb6fef7bd0957d1f984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4350827
Reviewed-by: Roman Arora <romanarora@chromium.org>
Commit-Queue: Roman Arora <romanarora@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120138}