Add a flag to disable canvas overdraw optimization
This change adds a feature flag that can be used to disable overdraw
optimizations in 2d canvas rendering contexts. The reason for adding
this flag is to be able to investigate hard to reproduce bugs in the
field to determine whether they are caused by this optimization.

Bug: 1423936
Change-Id: Ia99fb64ba7d8dd7fe8778d78b6e5a775d59cff65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4355668
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120179}