Have the Code Cache use the NIK when the HTTP Cache does
This CL updates the Code Cache to use the NIK when the
HTTP Cache does (which isn't only when
kSplitCacheByNetworkIsolationKey is enabled, as it once
was).

This is implemented behind it's own feature flag so that
it could be enabled via a slow rollout or as a killswitch.

For more context, see:
https://bugs.chromium.org/p/chromium/issues/detail?id=1213818#c30

Bug: 1213818
Change-Id: Iece67fb7f34269d1f1586de7854dc7526e50e7a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4303800
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120187}