[CacheStorage] Skip DeleteStorageKeyData call if no data to delete
This makes it so that we won't call
CacheStorageManager::DeleteStorageKeyData in the
CacheStorageContextImpl destructor unless we really need to. This
is particularly important because this code runs at shutdown and
has some recently identified issues related to CacheStorageManager
appearing to be destroyed on the wrong sequence.

Bug: 1419573
Change-Id: I412547babae1d134fbbe9cddb787feb84e7f1ec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4357087
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120203}