exo: Create wayland servers via FD
For go/securer-exo-ids, wayland servers will be created on-demand using
only an FD, and the responsibility of creating/binding that socket will
be taken over by other daemons.

In this CL we add the API to exo::wayland::Server() to create such a
server, as well as test infrastructure to work with it.

In future CLs we will migrate existing usage of StartAsync() and
Open(false) to use new new OpenFD() paths.

Bug: b:271515626
Change-Id: I5850181b73a2b8e3dd4316fb02f96be544f51f24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4328927
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/main@{#1120217}