ReadingList: Stop invoking MarkEntryUpdated() in SetReadStatusIfExists()
This CL removes the call to ReadingListEntry::MarkEntryUpdated() from
ReadingListModelImpl::SetReadStatusIfExists(). This is because:

1- ReadingListEntry::MarkEntryUpdated() should not be called when the
ReadingListEntry goes from the Unseen state to the Unread state. This is
not checked in ReadingListModelImpl::SetReadStatusIfExists().

2- ReadingListEntry::MarkEntryUpdated() is already called in
ReadingListEntry::SetRead() if the above condition is satisfied.

Therefore, calling ReadingListEntry::MarkEntryUpdated() from
ReadingListModelImpl::SetReadStatusIfExists() is either redundant or not
correct.

Bug: 1402196
Change-Id: I642d76936a79078197a54cb260c8dbb41cf02c8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4357098
Commit-Queue: Mahmoud Rashad <mmrashad@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120229}