webauthn: use a singleton type for mechanisms that don't need one.
Mechanisms use an `absl::variant` to express their type. For types like
`AddPhone`, there's no payload—all the information is expressed in the
type. But `base::StrongAlias` requires a payload type.

I used `bool` for this originally, which was silly. Additionally in
67b8eec267 I removed the explicit arguments to, e.g., `WindowsAPI`
because all the uses were just using `absl::holds_alternative` and thus
it wasn't needed. But then I used `==` in `IndexOfPriorityMechanism`
which will actually check that boolean payload.

It seems to be working per the tests, but I think that's because of
luck. This change fixes the original sin and uses `std::monostate`.

Bug: 1426245
Change-Id: Ie5643ab6989fcfc0f4547f15db85bcdfa9d76e6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4358591
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120238}